pyproject.toml
[tool.poetry]
name = "semantic-caching"
version = "0.1.0"
description = ""
authors = ["JimZer <jimi.vaubien@protonmail.com>"]
readme = "README.md"
packages = [{include = "semantic_caching"}]

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.0.202"
openai = "^0.27.8"
streamlit = "^1.23.1"
python-dotenv = "^1.0.0"
gptcache = "^0.1.32"
tiktoken = "^0.4.0"
rich = "^13.4.2"
torch = "^2.0.1"
typer = "^0.9.0"

[tool.poetry.scripts]
cli = "scripts.cli:app"
webapp = "scripts.app_runner:main"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
ruff = "^0.0.275"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
